% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/bump.R
\name{bump}
\alias{bump}
\title{Bump package version}
\usage{
bump(pkg = ".", component = 2L, format = "0.0-0")
}
\arguments{
\item{pkg}{package description, can be path or package name.  See
\code{\link{as.package}} for more information}

\item{component}{The position of the version to increase, starting at 1.}

\item{format}{The format of the version number, e.g., \code{"0.0-0"}}
}
\description{
This function increases the given component of the version number
in the DESCRIPTION file of a package.
}

