% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/bump-all.R
\name{bump_all}
\alias{bump_all}
\title{Bump package version and all dependent versions}
\usage{
bump_all(pkg, web = get_web_root(), component = 2L, format = "0.0-0")
}
\arguments{
\item{pkg}{Location of package}

\item{component}{The position of the version to increase, starting at 1.}
}
\description{
This function increases the given component of the version number
in the DESCRIPTION file of a package, and propagates this increase
to all reverse dependencies.
}

