% Generated by roxygen2 (4.1.1.9001): do not edit by hand
% Please edit documentation in R/deps-df.R
\name{deps_df}
\alias{deps_df}
\title{Return dependencies in a package web as data frame}
\usage{
deps_df(web = rpkgweb())
}
\description{
This function computes all dependencies (internal and external) in a
package web and returns the information as a data frame.
}

