% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/get-web-root.R
\name{get_web_root}
\alias{get_web_root}
\title{Finds the root of a package web}
\usage{
get_web_root(initial = getwd())
}
\arguments{
\item{initial}{Starting directory of the search}
}
\description{
Currently returns the argument which defaults to the working directory.
}

