% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/deps-df.R
\name{parse_deps}
\alias{parse_deps}
\title{Parse package dependency strings.}
\usage{
parse_deps(string)
}
\arguments{
\item{string}{to parse. Should look like \code{"R (>= 3.0), ggplot2"} etc.}
}
\value{
list of two character vectors: \code{name} package names,
  and \code{version} package versions. If version is not specified,
  it will be stored as NA.
}
\description{
Copy from devtools@dc61a9e79c44f4 to avoid loading devtools to parse
dependencies
}
\keyword{internal}

