% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/write-makefile.R
\name{write_makefile}
\alias{write_makefile}
\title{Read a package web}
\usage{
write_makefile(web = rpkgweb(), target_dir = NULL, lib_dir = NULL)
}
\arguments{
\item{web}{The web to construct the \code{Makefile} for}

\item{target_dir}{Target directory for \code{Makefile}, default: the web's root
directory}

\item{lib_dir}{Library directory for installation, default: User library}
}
\description{
This function reads a web of packages.
}

