% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/check-up.R
\name{check_up}
\alias{check_up}
\title{Check and install package if required}
\usage{
check_up(pkg_name, web = rpkgweb(), quiet = FALSE)
}
\arguments{
\item{pkg_name}{Name of the package, \emph{not} the path!}
}
\value{
\code{TRUE} if package has been updated, \code{FALSE} if the package
  is already up to date, or an error if there was a failure
}
\description{
If the currently installed version of a package differs
from the locally available version, this function uninstalls all reverse
dependencies, checks the package and installs it if all checks succeed.
}
\details{
TODO: devtools-like interface with pkg = "." and web = rpkgweb(pkg)

TODO: argument unload = !is_loaded(pkg)
}

