% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/check-up.R
\name{check_up}
\alias{check_up}
\title{Check and install package if required}
\usage{
check_up(pkg, web = rpkgweb())
}
\arguments{
\item{pkg}{Location of package}
}
\description{
If the currently installed version of a package differs
from the locally available version, this function checks the package
and installs it if all checks succeed.
}

