% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/makify.R
\name{makify}
\alias{makify}
\title{Construct a Makefile}
\usage{
makify(web = rpkgweb(), target_dir = NULL, lib_dir = NULL)
}
\arguments{
\item{web}{The web to construct the \code{Makefile} for}

\item{target_dir}{Target directory for \code{Makefile}, default: the web's root
directory}

\item{lib_dir}{Library directory for installation, default: User library}
}
\value{
A Makefile
}
\description{
This function constructs a \code{Makefile} for a package web.
}

