% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/makify.R
\name{makify}
\alias{makify}
\title{Construct a Makefile}
\usage{
makify(makefile, y)
}
\arguments{
\item{makefile}{A Makefile as created by \code{\link[MakefileR]{makefile}}}

\item{y}{The object to construct the \code{Makefile} for}
}
\value{
A Makefile
}
\description{
This function constructs a \code{Makefile} for various objects.
}

