% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/makify.R
\name{makify}
\alias{makify}
\title{Construct a Makefile}
\usage{
makify(x)
}
\arguments{
\item{x}{The object to construct the \code{Makefile} for}
}
\value{
A character vector that represents the contents of the generated
\code{Makefile}
}
\description{
This function constructs a \code{Makefile} for various objects.
}

