% Generated by roxygen2 (4.1.1.9001): do not edit by hand
% Please edit documentation in R/makify.R
\name{makify}
\alias{makify}
\title{Construct a Makefile}
\usage{
makify(web = rpkgweb())
}
\arguments{
\item{web}{The object to construct the \code{Makefile} for}
}
\value{
A Makefile
}
\description{
This function constructs a \code{Makefile} for various objects.
}

