% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/read-web.R
\name{read_web}
\alias{as.rpkgweb}
\alias{is.rpkgweb}
\alias{read_web}
\title{Read a package web}
\usage{
read_web(web = get_web_root())

is.rpkgweb(x)

as.rpkgweb(x)
}
\arguments{
\item{web}{A directory within a package web}
}
\description{
This function reads a web of packages.
}

