% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/read-web.R
\name{rpkgweb}
\alias{as.rpkgweb}
\alias{is.rpkgweb}
\alias{root_dir}
\alias{rpkgweb}
\title{Read a package web}
\usage{
rpkgweb(root_dir = get_web_root())

is.rpkgweb(x)

as.rpkgweb(x)

root_dir(x)
}
\arguments{
\item{root_dir}{The root directory of the package web}

\item{x}{An object}
}
\description{
An R package web is a collection of R packages stored in the same directory.
These function are related to the S3 class \code{"rpkgweb"}.
}
\details{
The \code{rpkgweb} function constructs an \code{"rpkgweb"} object
for a given directory.

\code{is.rpkgweb} checks if an object is of class \code{"rpkgweb"}.

All functions that operate on a package web will call \code{as.rpkgweb} on
  the input. Objects of class \code{"rpkgweb"} are passed through,
  \code{character} values are coerced using the \code{rpkgweb} function
  (i.e., they are interpreted as root path of the package web);
  other inputs will raise an error.

\code{root_dir} returns the root directory of a package web.
}

