% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/read-web.R
\name{rpkgweb}
\alias{as.rpkgweb}
\alias{is.rpkgweb}
\alias{root_dir}
\alias{rpkgweb}
\title{Read a package web}
\usage{
rpkgweb(root_dir = get_web_root())

is.rpkgweb(x)

as.rpkgweb(x)

root_dir(x)
}
\arguments{
\item{root_dir}{The root directory of the package web}

\item{x}{An object}
}
\description{
This function reads a web of packages.
}
\details{
\code{root_dir} returns the root directory of the web.
}

