% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/create-makefile.R
\name{write_makefile}
\alias{write_makefile}
\title{Read a package web}
\usage{
write_makefile(web = get_web_root())
}
\arguments{
\item{web}{A directory within a package web}
}
\description{
This function reads a web of packages.
}

